Refactor BigDecimal, implicit types used from return value
Rather than having to specify.
Also change constructor used to include a precision.

+++ src/main/scala/com/tomliddle/solution/timeanddistance/StraightLineTimeAndDistCalc.scala
@@ -22,7 +22,7 @@ trait StraightLineTimeAndDistCalc extends TimeAndDistCalc {
 		var a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
 			Math.sin(dLon / 2) * Math.sin(dLon / 2) * Math.cos(lat1) * Math.cos(lat2)
 		var c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
-		BigDecimal(R * c)
+		R * c
 	}
 
 	override def getDuration(stop1: Point, stop2: Point): Duration = {
